[ANDROSDK-1615-3] Remove period generators based on Calendar
