+++ core/src/androidTest/java/org/hisp/dhis/android/core/trackedentity/search/TrackedEntityInstanceLocalQueryHelperMockIntegrationShould.kt
@@ -35,6 +35,7 @@ import org.hisp.dhis.android.core.common.*
 import org.hisp.dhis.android.core.enrollment.EnrollmentStatus
 import org.hisp.dhis.android.core.event.EventStatus
 import org.hisp.dhis.android.core.organisationunit.OrganisationUnitMode
+import org.hisp.dhis.android.core.period.clock.internal.ClockProviderFactory
 import org.hisp.dhis.android.core.period.internal.CalendarProviderFactory
 import org.hisp.dhis.android.core.period.internal.ParentPeriodGeneratorImpl.Companion.create
 import org.hisp.dhis.android.core.trackedentity.internal.TrackedEntityInstanceStoreImpl
@@ -46,7 +47,8 @@ class TrackedEntityInstanceLocalQueryHelperMockIntegrationShould : BaseMockInteg
     private val trackedEntityInstanceStore = TrackedEntityInstanceStoreImpl(databaseAdapter)
 
     private val calendarProvider = CalendarProviderFactory.calendarProvider
-    private val periodHelper = DateFilterPeriodHelper(calendarProvider, create(calendarProvider))
+    private val clockProvider = ClockProviderFactory.clockProvider
+    private val periodHelper = DateFilterPeriodHelper(calendarProvider, create(clockProvider))
     private val localQueryHelper = TrackedEntityInstanceLocalQueryHelper(periodHelper)
 
     @Test
