+++ core/src/main/java/org/hisp/dhis/android/core/period/internal/ParentPeriodGenerator.kt
@@ -32,7 +32,7 @@ import org.hisp.dhis.android.core.period.Period
 import org.hisp.dhis.android.core.period.PeriodType
 import java.util.*
 
-interface ParentPeriodGenerator {
+internal interface ParentPeriodGenerator {
     fun generatePeriods(): List<Period>
     fun generatePeriods(periodType: PeriodType, endPeriods: Int): List<Period>
     fun generatePeriods(periodType: PeriodType, startPeriods: Int, endPeriods: Int): List<Period>
