+++ core/src/test/java/org/hisp/dhis/android/core/period/generator/internal/YearlyFinancialJulyPeriodGeneratorShould.kt
@@ -25,30 +25,44 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.period.internal.generator.internal
+package org.hisp.dhis.android.core.period.generator.internal
 
 import com.google.common.truth.Truth.assertThat
 import kotlinx.datetime.Clock
 import kotlinx.datetime.LocalDate
-import org.hisp.dhis.android.core.period.generator.internal.YearlyPeriodGeneratorFactory
+import org.hisp.dhis.android.core.period.clock.internal.fixed
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
 @RunWith(JUnit4::class)
-class YearlyFinancialNovemberPeriodGeneratorShould {
+class YearlyFinancialJulyPeriodGeneratorShould {
 
     @Test
     fun generate_periods_for_one_year() {
         val clock = Clock.fixed(LocalDate(2019, 2, 21))
 
-        val periods = YearlyPeriodGeneratorFactory.financialNov(clock).generatePeriods(-1, 0)
+        val periods = YearlyPeriodGeneratorFactory.financialJuly(clock).generatePeriods(-1, 0)
 
         assertThat(periods.size).isEqualTo(1)
         periods.first().let { period ->
-            assertThat(period.periodId).isEqualTo("2018Nov")
-            assertThat(period.startDate).isEqualTo(LocalDate(2017, 11, 1))
-            assertThat(period.endDate).isEqualTo(LocalDate(2018, 10, 31))
+            assertThat(period.periodId).isEqualTo("2017July")
+            assertThat(period.startDate).isEqualTo(LocalDate(2017, 7, 1))
+            assertThat(period.endDate).isEqualTo(LocalDate(2018, 6, 30))
+        }
+    }
+
+    @Test
+    fun generate_ending_period_on_june_30() {
+        val clock = Clock.fixed(LocalDate(2018, 6, 30))
+
+        val periods = YearlyPeriodGeneratorFactory.financialJuly(clock).generatePeriods(-1, 0)
+
+        assertThat(periods.size).isEqualTo(1)
+        periods.first().let { period ->
+            assertThat(period.periodId).isEqualTo("2016July")
+            assertThat(period.startDate).isEqualTo(LocalDate(2016, 7, 1))
+            assertThat(period.endDate).isEqualTo(LocalDate(2017, 6, 30))
         }
     }
 
@@ -56,34 +70,34 @@ class YearlyFinancialNovemberPeriodGeneratorShould {
     fun generate_periods_for_two_year() {
         val clock = Clock.fixed(LocalDate(2019, 2, 21))
 
-        val periods = YearlyPeriodGeneratorFactory.financialNov(clock).generatePeriods(-2, 0)
+        val periods = YearlyPeriodGeneratorFactory.financialJuly(clock).generatePeriods(-2, 0)
 
         assertThat(periods.size).isEqualTo(2)
 
-        assertThat(periods.map { it.periodId }).isEqualTo(listOf("2017Nov", "2018Nov"))
+        assertThat(periods.map { it.periodId }).isEqualTo(listOf("2016July", "2017July"))
     }
 
     @Test
     fun generate_period_id() {
-        val generator = YearlyPeriodGeneratorFactory.financialNov(Clock.System)
-        assertThat("2019Nov").isEqualTo(generator.generatePeriod(LocalDate(2019, 10, 30), 0).periodId)
-        assertThat("2020Nov").isEqualTo(generator.generatePeriod(LocalDate(2019, 11, 2), 0).periodId)
+        val generator = YearlyPeriodGeneratorFactory.financialJuly(Clock.System)
+        assertThat("2018July").isEqualTo(generator.generatePeriod(LocalDate(2019, 6, 30), 0).periodId)
+        assertThat("2019July").isEqualTo(generator.generatePeriod(LocalDate(2019, 7, 2), 0).periodId)
     }
 
     @Test
     fun generate_period_id_with_offset() {
-        val generator = YearlyPeriodGeneratorFactory.financialNov(Clock.System)
-        assertThat("2020Nov").isEqualTo(generator.generatePeriod(LocalDate(2019, 10, 30), 1).periodId)
-        assertThat("2019Nov").isEqualTo(generator.generatePeriod(LocalDate(2019, 11, 2), -1).periodId)
+        val generator = YearlyPeriodGeneratorFactory.financialJuly(Clock.System)
+        assertThat("2019July").isEqualTo(generator.generatePeriod(LocalDate(2019, 6, 30), 1).periodId)
+        assertThat("2018July").isEqualTo(generator.generatePeriod(LocalDate(2019, 7, 2), -1).periodId)
     }
 
     @Test
     fun generate_periods_in_this_year() {
         val clock = Clock.fixed(LocalDate(2020, 9, 29))
 
-        val generator = YearlyPeriodGeneratorFactory.financialNov(clock)
+        val generator = YearlyPeriodGeneratorFactory.financialJuly(clock)
         val periods = generator.generatePeriodsInYear(0)
         assertThat(periods.size).isEqualTo(1)
-        assertThat(periods.first().periodId).isEqualTo("2020Nov")
+        assertThat(periods.first().periodId).isEqualTo("2020July")
     }
 }
